+++ src/main/java/top/girlkisser/cygnus/foundation/space/PlanetaryDangerIndex.java
@@ -1,0 +1,10 @@
+package top.girlkisser.cygnus.foundation.space;
+
+public enum PlanetaryDangerIndex
+{
+	HABITABLE,
+	SAFE,
+	EUCLID,
+	LETHAL,
+	NON_LANDABLE,
+}